+++ build.gradle
@@ -47,7 +47,7 @@ dependencies {
 	implementation 'io.projectreactor:reactor-core:3.6.5' // 예시로 최신 버전
 	implementation 'org.springframework.boot:spring-boot-starter-webflux'
 	implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.2'	//Swagger
-	implementation 'io.reactivex.rxjava3:rxjava:3.1.6' // rxjava
+	implementation 'io.reactivex.rxjava3:rxjava:3.1.8' // rxjava
 	implementation 'org.springframework.boot:spring-boot-starter-mail'
 	implementation 'org.springframework.boot:spring-boot-starter-websocket'
